Simplify Card class to fix hellacious bug
The logic is some of the Card constructors, notibly build_from_value and
build_from_face_suit_values was overly complicated and under tested. The
result was a bug where the Card#value did not match the corresponding
suit and face value.

This commit changes the calculate of value (simplifies it a lot). The
value should be internal to the Card class but this is a breaking change
if anyone was depending on it to be stable.